+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2023-07-17
+- Added minimal hypothesis length validation for new `Submission` instances.
+
 ## [0.7.0] - 2023-07-14
 - Updated `Submission.file_format` default value to `XML_FILE`.
 - Updated `SubmissionForm` and made `file_format` widget readonly. This ensures that via the UI only XML files can be submitted. [[#101](https://github.com/AppraiseDev/OCELoT/issues/101)]